refact: transactions standardized as contracts
