+++ chains/transactionManager.js
@@ -1,17 +1,12 @@
-// src/chains/transactionManager.js
-import { addAgreement as ethAddAgreement } from "./ethereum/transactions";
-import { addAgreement as solAddAgreement } from "./solana/transactions";
+import solTransactions from "./solana/transactions"
+import ethTransactions from "./ethereum/transactions"
 
 class TransactionManager {
   constructor() {
     this.chains = {
-      ethereum: {
-        addAgreement: ethAddAgreement
-      },
-      solana: {
-        addAgreement: solAddAgreement
-      },
-    };    
+      ethereum: ethTransactions,
+      solana: solTransactions,
+    }
   }
 
   async addAgreement(chain, ...args) {
